Small refactors (#289)
