+++ README.md
@@ -201,7 +201,7 @@ $ image_id: docker images --- --column 3 --header-lines 1 --delimiter '\s\s+'
 The supported parameters are:
 - `--prevent-extra` *(experimental)*: limits the user to select one of the suggestions;
 - `--column <number>`: extracts a single column from the selected result;
-- `--delimiter <regex>`: delimits columns;
+- `--delimiter <regex>`: delimits columns + forwarded option to `fzf`;
 - `--multi`: forwarded option to `fzf`;
 - `--header-lines <number>`: forwarded option to `fzf`;
