--- src/cmds/query.rs
@@ -1,8 +1,0 @@
-use crate::cmds;
-use crate::cmds::core::Variant;
-use crate::option::Config;
-use std::error::Error;
-
-pub fn main(query: String, config: Config) -> Result<(), Box<dyn Error>> {
-    cmds::core::main(Variant::Query(query), config, true)
-}
