+++ src/filesystem.rs
@@ -1,4 +1,4 @@
-use crate::option::Config;
+use crate::structures::option::Config;
 use std::fs;
 use std::fs::File;
 use std::io::{self, BufRead, BufReader, Lines};
