+++ src/flows/search.rs
@@ -1,5 +1,5 @@
 use super::aux;
-use crate::option::Config;
+use crate::structures::option::Config;
 use std::error::Error;
 
 pub fn main(_query: String, _config: Config) -> Result<(), Box<dyn Error>> {
