+++ src/main.rs
@@ -1,20 +1,19 @@
 #[macro_use]
 extern crate lazy_static;
 
-mod cheat;
-mod cmds;
 mod display;
 mod filesystem;
-mod fnv;
+mod flows;
 mod fzf;
 mod git;
 mod handler;
-mod option;
+mod parser;
+mod structures;
 mod terminal;
 mod welcome;
 
 use std::error::Error;
 
 fn main() -> Result<(), Box<dyn Error>> {
-    handler::handle_config(option::config_from_env())
+    handler::handle_config(structures::option::config_from_env())
 }
