+++ src/parser.rs
@@ -1,52 +1,13 @@
 use crate::display;
 use crate::filesystem;
-use crate::fnv::HashLine;
-use crate::option::Config;
+use crate::structures::cheat::{SuggestionOpts, SuggestionType, VariableMap};
+use crate::structures::fnv::HashLine;
+use crate::structures::option::Config;
 use crate::welcome;
 use regex::Regex;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::fs;
 use std::io::Write;
-
-#[derive(Debug, PartialEq)]
-pub struct SuggestionOpts {
-    pub header_lines: u8,
-    pub column: Option<u8>,
-    pub delimiter: Option<String>,
-    pub suggestion_type: SuggestionType,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum SuggestionType {
-    /// fzf will not print any suggestions
-    Disabled,
-    /// fzf will only select one of the suggestions
-    SingleSelection,
-    /// fzf will select multiple suggestions
-    MultipleSelections,
-    /// fzf will select one of the suggestions or use the query
-    SingleRecommendation,
-    /// initial snippet selection
-    SnippetSelection,
-}
-
-pub type Suggestion = (String, Option<SuggestionOpts>);
-
-pub struct VariableMap(HashMap<u64, Suggestion>);
-
-impl VariableMap {
-    pub fn new() -> Self {
-        Self(HashMap::new())
-    }
-
-    pub fn insert(&mut self, tags: &str, variable: &str, value: Suggestion) -> Option<Suggestion> {
-        self.0.insert(gen_key(tags, variable), value)
-    }
-
-    pub fn get(&self, tags: &str, variable: &str) -> Option<&Suggestion> {
-        self.0.get(&gen_key(tags, variable))
-    }
-}
 
 fn remove_quotes(txt: &str) -> String {
     txt.replace('"', "").replace('\'', "")
@@ -118,15 +79,11 @@ fn write_cmd(
     }
 }
 
-fn gen_key(tags: &str, variable: &str) -> u64 {
-    format!("{};{}", tags, variable).hash_line()
-}
-
 fn read_file(
     path: &str,
     variables: &mut VariableMap,
+    visited_lines: &mut HashSet<u64>,
     stdin: &mut std::process::ChildStdin,
-    set: &mut HashSet<u64>,
 ) -> bool {
     let mut tags = String::from("");
     let mut comment = String::from("");
@@ -143,10 +100,10 @@ fn read_file(
 
             let line = l.unwrap();
             let hash = line.hash_line();
-            if set.contains(&hash) {
+            if visited_lines.contains(&hash) {
                 continue;
             }
-            set.insert(hash);
+            visited_lines.insert(hash);
 
             // blank
             if line.is_empty() {
@@ -203,9 +160,9 @@ fn read_file(
 pub fn read_all(config: &Config, stdin: &mut std::process::ChildStdin) -> VariableMap {
     let mut variables = VariableMap::new();
     let mut found_something = false;
+    let mut visited_lines = HashSet::new();
     let paths = filesystem::cheat_paths(config);
     let folders = paths.split(':');
-    let mut set = HashSet::new();
 
     for folder in folders {
         if let Ok(paths) = fs::read_dir(folder) {
@@ -213,7 +170,7 @@ pub fn read_all(config: &Config, stdin: &mut std::process::ChildStdin) -> Variab
                 let path = path.unwrap().path();
                 let path_str = path.to_str().unwrap();
                 if path_str.ends_with(".cheat")
-                    && read_file(path_str, &mut variables, stdin, &mut set)
+                    && read_file(path_str, &mut variables, &mut visited_lines, stdin)
                     && !found_something
                 {
                     found_something = true;
@@ -257,10 +214,9 @@ mod tests {
         let mut variables = VariableMap::new();
         let mut child = Command::new("cat").stdin(Stdio::piped()).spawn().unwrap();
         let child_stdin = child.stdin.as_mut().unwrap();
-        let mut set: HashSet<u64> = HashSet::new();
-        read_file(path, &mut variables, child_stdin, &mut set);
-        let mut result = VariableMap::new();
-        let suggestion = (
+        let mut visited_lines: HashSet<u64> = HashSet::new();
+        read_file(path, &mut variables, &mut visited_lines, child_stdin);
+        let expected_suggestion = (
             r#" echo -e "$(whoami)\nroot" "#.to_string(),
             Some(SuggestionOpts {
                 header_lines: 0,
@@ -269,19 +225,7 @@ mod tests {
                 suggestion_type: SuggestionType::SingleRecommendation,
             }),
         );
-        result.insert("ssh", "user", suggestion);
-        let actual_suggestion = result.get("ssh", "user");
-        assert_eq!(
-            Some(&(
-                r#" echo -e "$(whoami)\nroot" "#.to_string(),
-                Some(SuggestionOpts {
-                    header_lines: 0,
-                    column: None,
-                    delimiter: None,
-                    suggestion_type: SuggestionType::SingleRecommendation,
-                }),
-            )),
-            actual_suggestion
-        );
+        let actual_suggestion = variables.get("ssh", "user");
+        assert_eq!(Some(&expected_suggestion), actual_suggestion);
     }
 }
