+++ src/structures/mod.rs
@@ -1,0 +1,4 @@
+pub mod cheat;
+pub mod fnv;
+pub mod fzf;
+pub mod option;
